chore: reorganize eth/storage files
