+++ src/bin/historic_events_processor.rs
@@ -6,10 +6,10 @@ use chrono::TimeZone;
 use chrono::Timelike;
 use indicatif::ProgressBar;
 use rocksdb::properties::ESTIMATE_NUM_KEYS;
-use stratus::eth::storage::rocks::rocks_state::RocksStorageState;
-use stratus::eth::storage::rocks::types::BlockRocksdb;
-use stratus::eth::storage::rocks::types::TransactionMinedRocksdb;
-use stratus::eth::storage::rocks::types::UnixTimeRocksdb;
+use stratus::eth::storage::permanent::rocks::types::BlockRocksdb;
+use stratus::eth::storage::permanent::rocks::types::TransactionMinedRocksdb;
+use stratus::eth::storage::permanent::rocks::types::UnixTimeRocksdb;
+use stratus::eth::storage::permanent::rocks::RocksStorageState;
 use stratus::ledger::events::transaction_to_events;
 use stratus::ledger::events::AccountTransfers;
 use stratus::ledger::events::Event;
