+++ src/bin/importer_offline.rs
@@ -17,14 +17,14 @@ use futures::StreamExt;
 use itertools::Itertools;
 use stratus::config::ImporterOfflineConfig;
 use stratus::eth::executor::Executor;
+use stratus::eth::external_rpc::ExternalBlockWithReceipts;
+use stratus::eth::external_rpc::ExternalRpcStorage;
 use stratus::eth::miner::Miner;
 use stratus::eth::miner::MinerMode;
 use stratus::eth::primitives::Block;
 use stratus::eth::primitives::BlockNumber;
 use stratus::eth::primitives::ExternalReceipts;
 use stratus::eth::primitives::ExternalTransaction;
-use stratus::eth::storage::ExternalBlockWithReceipts;
-use stratus::eth::storage::ExternalRpcStorage;
 use stratus::ext::spawn_named;
 use stratus::ext::spawn_thread;
 use stratus::log_and_err;
