+++ src/bin/rpc_downloader.rs
@@ -12,10 +12,10 @@ use futures::StreamExt;
 use itertools::Itertools;
 use serde::Deserialize;
 use stratus::config::RpcDownloaderConfig;
+use stratus::eth::external_rpc::ExternalRpcStorage;
 use stratus::eth::primitives::Address;
 use stratus::eth::primitives::BlockNumber;
 use stratus::eth::primitives::Hash;
-use stratus::eth::storage::ExternalRpcStorage;
 use stratus::ext::not;
 use stratus::infra::BlockchainClient;
 use stratus::utils::DropTimer;
