+++ src/eth/external_rpc/sql/insert_external_balance.sql
