+++ src/eth/external_rpc/sql/insert_external_block_and_receipts.sql
