+++ src/eth/external_rpc/sql/select_external_balances.sql
