+++ src/eth/external_rpc/sql/select_external_blocks_and_receipts_in_range.sql
