+++ src/eth/external_rpc/sql/select_external_blocks_in_range.sql
