+++ src/eth/external_rpc/sql/select_max_external_block_in_range.sql
