+++ src/eth/mod.rs
@@ -1,5 +1,6 @@
 pub mod codegen;
 pub mod executor;
+pub mod external_rpc;
 pub mod follower;
 pub mod miner;
 pub mod primitives;
