--- src/eth/storage/inmemory/mod.rs
@@ -1,8 +1,0 @@
-mod inmemory_history;
-mod inmemory_permanent;
-mod inmemory_temporary;
-
-pub use inmemory_history::InMemoryHistory;
-pub use inmemory_permanent::InMemoryPermanentStorage;
-pub use inmemory_permanent::InMemoryPermanentStorageState;
-pub use inmemory_temporary::InMemoryTemporaryStorage;
