+++ src/eth/storage/permanent/redis.rs
