+++ src/eth/storage/permanent/rocks/mod.rs
@@ -1,10 +1,13 @@
 //! RocksDB layers (top-to-bottom): permanent -> state -> rest.
 
+pub use rocks_permanent::RocksPermanentStorage;
+pub use rocks_state::RocksStorageState;
+
 /// Exposed API.
-pub mod rocks_permanent;
+mod rocks_permanent;
 
 /// State handler for DB and column families.
-pub mod rocks_state;
+mod rocks_state;
 
 /// CFs versionated by value variant.
 mod cf_versions;
