+++ src/eth/storage/permanent/rocks/rocks_batch_writer.rs
