+++ src/eth/storage/permanent/rocks/types/bytes.rs
