+++ src/eth/storage/permanent/rocks/types/difficulty.rs
