+++ src/eth/storage/permanent/rocks/types/execution_result.rs
