+++ src/eth/storage/permanent/rocks/types/hash.rs
