+++ src/eth/storage/permanent/rocks/types/log.rs
