+++ src/eth/storage/permanent/rocks/types/log_mined.rs
