+++ src/eth/storage/permanent/rocks/types/slot.rs
