+++ src/eth/storage/permanent/rocks/types/transaction_mined.rs
