--- src/eth/storage/postgres_external_rpc/mod.rs
@@ -1,5 +1,0 @@
-#[allow(clippy::module_inception)]
-mod postgres_external_rpc;
-
-pub use postgres_external_rpc::PostgresExternalRpcStorage;
-pub use postgres_external_rpc::PostgresExternalRpcStorageConfig;
