--- src/eth/storage/redis/mod.rs
@@ -1,3 +1,0 @@
-mod redis_permanent;
-
-pub use redis_permanent::RedisPermanentStorage;
